+++ table-api/table_entry.go
@@ -1,6 +1,12 @@
 package tableapi
 
-import "fmt"
+import "errors"
+
+const (
+	displayValueKey = "display_value"
+	valueKey        = "value"
+	linkKey         = "link"
+)
 
 // TableEntry represents a single Service-Now table entry.
 type TableEntry map[string]interface{}
@@ -17,8 +23,6 @@ func (tE TableEntry) Set(key string, value interface{}) {
 	tE[key] = value
 }
 
-// Deprecated: deprecated since v{unreleased}. Please use Get instead.
-//
 // Value retrieves a TableValue instance for the given key.
 func (tE TableEntry) Value(key string) *TableValue {
 	value, exists := tE[key]
@@ -30,11 +34,54 @@ func (tE TableEntry) Value(key string) *TableValue {
 
 	switch v := value.(type) {
 	case map[string]interface{}:
-		trueVal = v["value"]
+		trueVal = v[valueKey]
 	case interface{}:
 		trueVal = v
 	}
 	return &TableValue{value: trueVal}
+}
+
+// DisplayValue retrieves a TableValue instance for the given key.
+func (tE TableEntry) DisplayValue(key string) *TableValue {
+	value, exists := tE[key]
+	if !exists {
+		return nil
+	}
+
+	var trueVal interface{}
+
+	switch v := value.(type) {
+	case map[string]interface{}:
+		trueVal = v[displayValueKey]
+	case interface{}:
+		trueVal = v
+	}
+	return &TableValue{value: trueVal}
+}
+
+// Link retrieves a String instance for the given key.
+func (tE TableEntry) Link(key string) (*string, error) {
+	value, exists := tE[key]
+	if !exists {
+		return nil, nil
+	}
+
+	var trueVal string
+
+	switch v := value.(type) {
+	case map[string]interface{}:
+		val, ok := v[linkKey]
+		if !ok {
+			return nil, nil
+		}
+		trueVal, ok = val.(string)
+		if !ok {
+			return nil, errors.New("link is not string")
+		}
+	default:
+		return nil, nil
+	}
+	return &trueVal, nil
 }
 
 // Keys returns a slice of all field names stored in the table entry.
@@ -49,63 +96,4 @@ func (tE TableEntry) Keys() []string {
 // Len returns the number of fields in the table entry.
 func (tE TableEntry) Len() int {
 	return len(tE)
-}
-
-// Get retrieves a RecordElementModel from the table entry based on the provided key.
-//
-// If the key is found, it processes the stored value and extracts necessary
-// metadata such as display value, value, and links.
-func (tE TableEntry) Get(key string) (*RecordElementModel, error) {
-	value, ok := tE[key]
-	if !ok {
-		return nil, nil
-	}
-
-	model := NewRecordElement()
-
-	switch v := value.(type) {
-	case map[string]interface{}:
-		if err := model.SetDisplayValue(v[displayValueKey]); err != nil {
-			return nil, err
-		}
-		if err := model.SetValue(v[valueKey]); err != nil {
-			return nil, err
-		}
-		if val, ok := v[linkKey]; ok {
-			if link, ok := val.(*string); ok {
-				if err := model.setLink(link); err != nil {
-					return nil, err
-				}
-			}
-		}
-	case interface{}:
-		if err := model.SetValue(v); err != nil {
-			return nil, err
-		}
-	default:
-		return nil, fmt.Errorf("unsupported type: %T", v)
-	}
-	return model, nil
-}
-
-// SetElement assigns a RecordElement to the specified key.
-func (tE TableEntry) SetElement(key string, element RecordElement) error {
-	tE[key] = element
-
-	return nil
-}
-
-// SetValue assigns a simple value to the specified key using a RecordElement wrapper.
-func (tE TableEntry) SetValue(key string, value any) error {
-	model := NewRecordElement()
-	if err := model.SetValue(value); err != nil {
-		return err
-	}
-	return tE.SetElement(key, model)
-}
-
-// HasAttribute checks if the specified key exists in the table entry.
-func (tE TableEntry) HasAttribute(key string) bool {
-	_, ok := tE[key]
-	return ok
 }