+++ changelog.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Batch API support ([Issue #49](https://github.com/michaeldcanady/servicenow-sdk-go/issues/49))
+
 ### Changed
 
 ### Deprecated
@@ -27,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - `core.RequestInformation.ToRequestWithContext` not including headers when converting to request ([Issue #107](https://github.com/michaeldcanady/servicenow-sdk-go/issues/107))
+- unable to access `display_value` and `link` with new implementation `tableapi.RecordElementModel` ([Issue #204](https://github.com/michaeldcanady/servicenow-sdk-go/issues/204))
 
 ## [1.6.0] - (08/03/2024)